Improve converting line endings from CR to LF
Add a new flag, `-ll` to the zip command used in fix_ascii. This
effectively changes the line endings of a file to LF without clobbering
binary files. If a binary file *is* changed, the command will issue a
warning.

Separately, also renamed the temporary file from "foo.zip" to a new
filename based off of `mktemp`, to avoid clobbering any current folders
/ files.