Update search to use `ag` if available
Ag ("the silver searcher") is an alternative to grep and ack that, apart
from being 50% and 33% faster to type respectively is significantly
faster than ack. It also ignores your .gitignore files (and .hgignore if
that's your thing) unless you tell it otherwise by providing the -U
flag.

In summary, it's faster and better than ack and grep. Only problem is
it doesn't come standard, so now this function checks if it's available.

http://geoff.greer.fm/2011/12/27/the-silver-searcher-better-than-ack/